user authorization was performed

+++ src/lesson35/service/HotelService.java
@@ -3,7 +3,10 @@ package lesson35.service;
 import lesson35.dao.HotelDAO;
 import lesson35.exception.BadRequestException;
 import lesson35.model.Hotel;
+import lesson35.model.User;
+import lesson35.model.UserType;
 
+import static lesson35.service.UserService.loggedInUser;
 import static lesson35.service.ValidationUtils.checkContentAndNull;
 
 
@@ -12,6 +15,11 @@ public class HotelService {
     public final HotelDAO hotelDAO = new HotelDAO();
 
     public Hotel findHotelByCity(String city) throws Exception {
+        User user = loggedInUser;
+        if (user.getUserType() == null) {
+            throw new BadRequestException("error: //.");
+        }
+
         for (Hotel hotel : hotelDAO.readAll()) {
             if (hotel.getCity().equals(city)) {
                 return hotel;
@@ -21,6 +29,11 @@ public class HotelService {
     }
 
     public Hotel findHotelByName(String name) throws Exception {
+        User user = loggedInUser;
+        if (user.getUserType() != UserType.ADMIN || user.getUserType() != UserType.USER) {
+            throw new BadRequestException("Error...");
+        }
+
         for (Hotel hotel : hotelDAO.readAll()) {
             if (hotel.getName().equals(name)) {
                 return hotel;
@@ -30,11 +43,21 @@ public class HotelService {
     }
 
     public void addHotel(Hotel hotel) throws Exception {
+        User user = loggedInUser;
+        if (user.getUserType() != UserType.ADMIN || user.getUserType() != UserType.USER) {
+            throw new BadRequestException("Error...");
+        }
+
         validateHotel(hotel);
         hotelDAO.addObjectToFile(hotel);
     }
 
     public void deleteHotel(long idHotel) throws Exception {
+        User user = loggedInUser;
+        if (user.getUserType() != UserType.ADMIN) {
+            throw new BadRequestException("error: the user is not an admin.");
+        }
+
         hotelDAO.findObject(idHotel);
         hotelDAO.deleteObjectFromFile(idHotel);
     }
@@ -44,4 +67,5 @@ public class HotelService {
             throw new BadRequestException("Error, the entered data is incomplete, fill in each specified field.");
         }
     }
+
 }
